safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/eval-scripts-setInterval-blocked-expected.txt

CONSOLE MESSAGE: Refused to evaluate script because it violates the following Content Security Policy directive: "script-src 'unsafe-inline'".

ALERT: PASS
